Add files via upload
Dell analytics case study is based on a different dataset related to the Dell company and with the use of different if functions we have focused on solving a few questions related to the dataset. Starting from calculating the sales target and the bonus amount for the employees to calculating the total experience of the employees at Dell is analyzed using the if functions.